requirements.txt
Pyrogram
TgCrypto
hurry.filesize
python-dotenv
dropbox